backend/requirements.txt
FastAPI==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
neo4j==5.14.0
google-generativeai==0.3.0
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2
aiofiles==23.2.1
slowapi==0.1.9
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
